+++ Craftsman/Services/ApiScaffoldingService.cs
@@ -152,6 +152,7 @@ public class ApiScaffoldingService
         _mediator.Send(new DomainEventBuilder.DomainEventBuilderCommand());
 
         //services
+        _mediator.Send(new UnitTestUtilsBuilder.Command());
         _mediator.Send(new UnitOfWorkBuilder.UnitOfWorkBuilderCommand(template.DbContext.ContextName));
         _mediator.Send(new IBoundaryServiceInterfaceBuilder.IBoundaryServiceInterfaceBuilderCommand());
         _mediator.Send(new GenericRepositoryBuilder.GenericRepositoryBuilderCommand(template.DbContext.ContextName));